DP-2162 Handle JSON errors at root
This change adds handling of JSON validation errors w/o a path.

E.g. if given a schema requiring the keys `a` and `b`, an object like
`{"a": 123}` will give a validation error of `"'b' is a required property"`,
but would not have a path.

`path` and `relative_path` are empty for validation errors at root:
https://python-jsonschema.readthedocs.io/en/latest/errors/#jsonschema.exceptions.ValidationError.relative_path