DP-2231 Mock less in CSV validator tests
Make less use of mocking in the CSV validator tests and test against an
actual moto S3 test bucket instead to cover a larger part of the data
flow.

This should make it easier to verify that gzipped files are handled
properly later (a failing test is added for that case now).