Rely on pre installed runtime libraries
Don't deploy libraries that are already pre-installed in the Lambda
runtime in order to save space. This shaves off ~60 MB on the package
deploy size (unzipped), making it deployable again.

Also lock the library versions to match those found in the Lambda
runtime in order to make our local testing environment match it as
closely as possible.